perf(server): load rdeploy-blob challenge data asynchronously
Do an asynchronous `readFile` and use `JSON.parse` when reading in
rDeploy's config.json instead of `require`ing it. Add a before hook to
tests regarding challenge submission to wait for challenge data to load
before trying to submit flags to that challenge - currently implemented
via polling on a 100ms interval until data is available.