fix(client): defer attempt to scroll to own team until it is loaded
If the user's team is not on the current scoreboard page, defer the
attempt to scroll it into view until the data for the scoreboard page
the user is on is actually loaded.

A possible optimization for this is to scroll to the element where the
user's team "would" be located on the scoreboard before the data has
loaded, however that would require us to reuse the host elements in the
table when the scoreboard page changes.