fix(client): make account deletion dialog update when you type
Event handler for controlled `input` element in the form in the account
deletion dialog for confirming team name used the wrong event (onChange
only fires when the user is done editing rather than as they are
typing). Fix this so that the "Confirm" button will reactivate as soon
as the correct team name is typed in the dialog.