feat(client): make logout button a modal dialog
The logout route has been replaced with a logout button that displays a
modal confirming logout.

Refactorings: Header made to use preact-router's `Match` element to
detect when a link is active. This removes the need for `App` to keep
track of the current path.

BUGS: due to `App` no longer updating on route changes, the header items
do not update when the user logs in or out.