fix(server): add error message for malformed data uris
The current implementation of the /api/v1/admin/upload endpoint does not distinguish between errors
from the upload provider, and errors from a malformed data URI provided by the user. This commit
adds a separate response message for malformed user input that returns a status code of 400 instead
of 500.

fix #300